# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SolveX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault/withdraw:
    post:
      operationId: withdrawVault
      tags: [vault]
      summary: Withdraw available vault funds (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Withdrawal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawResult"
...
  /solutions:
    post:
      operationId: submitSolution
      tags: [solutions]
      summary: Submit a solution (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionInput"
      responses:
        "201":
          description: Solution submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
  /solutions/problem/{problemId}:
    get:
      operationId: getSolutionsByProblem
      tags: [solutions]
      summary: Get solutions for a problem
      parameters:
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Solutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Solution"
  /solutions/{id}/verify:
    post:
      operationId: verifySolution
      tags: [solutions]
      summary: AI-verify a solution (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"

  # ── Escrow ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WithdrawInput:
      type: object
      required: [entryIds, withdrawalAddress]
      properties:
        entryIds:
          type: array
          items:
            type: string
        withdrawalAddress:
          type: string
    WithdrawResult:
      type: object
      required: [success, count]
      properties:
        success:
          type: boolean
        count:
          type: integer
        totalAmount:
          type: string
...
    Solution:
      type: object
      required: [id, problemId, solverId, content, status]
      properties:
        id:
          type: integer
        problemId:
          type: integer
        solverId:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [pending, verifying, approved, rejected]
        verificationScore:
          type: ["string", "null"]
        verificationNotes:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SolutionInput:
      type: object
      required: [problemId, content]
      properties:
        problemId:
          type: integer
        content:
          type: string
          minLength: 20
    VerificationResult:
      type: object
      required: [approved, score]
      properties:
        approved:
          type: boolean
        score:
          type: number
        notes:
          type: string
...